+++ src/app/modules/map/component/map-dialog/map-dialog.component.spec.ts
@@ -35,7 +35,8 @@ describe('MapDialogComponent', () => {
         {
           provide: MAT_DIALOG_DATA,
           useValue: {
-            item: {}
+            item: {},
+            settings: {}
           }
         },
         { provide: ElectronProvider, useClass: ElectronProviderFake }