+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 0.6.21 (2020-04-12)
 
 - add focus game after closing all dialogs (#606)
-- add move poe overlay to top after focusing poe (#608)
+- add move poe overlay to top after focusing poe (#608) 
 - update keyboard support to be always enabled
+- remove quality min/ max restriction (#611)
 
 ## 0.6.20 (2020-04-10)